SP-359 refactor: 성공 응답 "seccess" 통일
