+++ src/main/java/com/ludo/study/studymatchingplatform/auth/kakao/controller/KakaoLoginController.java
@@ -14,7 +14,7 @@ import com.ludo.study.studymatchingplatform.auth.common.provider.CookieProvider;
 import com.ludo.study.studymatchingplatform.auth.common.provider.JwtTokenProvider;
 import com.ludo.study.studymatchingplatform.auth.kakao.service.KakaoLoginService;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
-import com.ludo.study.studymatchingplatform.study.controller.dto.BaseApiResponse;
+import com.ludo.study.studymatchingplatform.study.controller.dto.response.BaseApiResponse;
 import com.ludo.study.studymatchingplatform.user.domain.user.Social;
 import com.ludo.study.studymatchingplatform.user.domain.user.User;
 import com.ludo.study.studymatchingplatform.user.service.dto.response.UserResponse;
@@ -54,7 +54,7 @@ public class KakaoLoginController {
 		final UserResponse userResponse = UserResponse.from(user);
 		cookieProvider.setAuthCookie(accessToken, response);
 		response.sendRedirect("https://local.ludoapi.store:3000");
-		return ResponseEntity.ok(BaseApiResponse.success("로그인 성공", userResponse));
+		return ResponseEntity.ok(BaseApiResponse.success(userResponse));
 	}
 
 }
