+++ src/main/java/com/ludo/study/studymatchingplatform/study/controller/dto/response/BaseApiResponse.java
@@ -11,8 +11,8 @@ public class BaseApiResponse<T> {
 	private final String message;
 	private final T data;
 
-	public static <T> BaseApiResponse<T> success(String message, T data) {
-		return new BaseApiResponse<>(true, message, data);
+	public static <T> BaseApiResponse<T> success(T data) {
+		return new BaseApiResponse<>(true, "success", data);
 	}
 
 	public static <T> BaseApiResponse<T> fail(String message, T data) {
