+++ src/main/java/com/ludo/study/studymatchingplatform/user/controller/UserController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ludo.study.studymatchingplatform.auth.common.AuthUser;
 import com.ludo.study.studymatchingplatform.auth.common.Redirection;
 import com.ludo.study.studymatchingplatform.auth.common.provider.CookieProvider;
-import com.ludo.study.studymatchingplatform.study.controller.dto.BaseApiResponse;
+import com.ludo.study.studymatchingplatform.study.controller.dto.response.BaseApiResponse;
 import com.ludo.study.studymatchingplatform.user.domain.user.User;
 import com.ludo.study.studymatchingplatform.user.service.UserService;
 import com.ludo.study.studymatchingplatform.user.service.dto.response.UserResponse;
@@ -41,7 +41,7 @@ public class UserController {
 	@GetMapping("/users/me")
 	public ResponseEntity<BaseApiResponse<UserResponse>> fetchMe(@AuthUser final User user) {
 		final UserResponse response = UserResponse.from(user);
-		return ResponseEntity.ok(BaseApiResponse.success("사용자 조회 성공", response));
+		return ResponseEntity.ok(BaseApiResponse.success(response));
 	}
 
 }
